pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "opslib-contrib"
dynamic = ["version", "description"]
dependencies = [
    "pyopslib",
    "requests",
    "pyyaml",
]
